requirements.txt
google-api-python-client==2.161.0
prometheus_client==0.21.1
numpy==1.26.4
pandas==2.2.3
yt-dlp==2025.1.15
Flask==2.3.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
flask-restx==0.5.1
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
prometheus-flask-exporter==0.23.1
psycopg2-binary==2.9.10
celery-prometheus-exporter==1.7.0
selenium==4.27.1
beautifulsoup4==4.11.1
webdriver-manager==4.0.2
celery==5.4.0
redis==5.2.1
flower==2.0.1
pydub==0.25.1